+++ src/app/actions.ts
@@ -48,33 +48,3 @@ export async function getUserData() {
 
   return userData;
 }
-
-/*
-Paperless
-|  _ \ __ _ _ __   ___ _ __| | ___ ___ ___
-| |_) / _` | '_ \ / _ | '__| |/ _ / __/ __|
-|  __| (_| | |_) |  __| |  | |  __\__ \__ \
-|_|   \__,_| .__/ \___|_|  |_|\___|___|___/
-           |_|
-*/
-
-export async function getPaperlessDocuments(query: string) {
-  const userData = await getUserData();
-
-  if (!query || query == "null" || query.length < 3 || !userData) return null;
-
-  const response = await fetch(
-    `${userData.paperlessURL}/api/documents/?query=${query}&page=1&page_size=10&truncate_content=true`,
-    {
-      method: "GET",
-      headers: {
-        "Content-Type": "application/json",
-        Authorization: `Token ${userData.paperlessToken}`,
-      },
-    },
-  );
-
-  const data = (await response.json()) as PaperlessDocumentsType;
-
-  return data;
-}
